requirements.txt
alembic==1.13.1
bcrypt==4.1.2
fastapi==0.110.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_settings==2.2.1
pyjwt==2.8.0
pytz==2024.1
requests==2.31.0
sqlalchemy==2.0.27
uvicorn==0.27.1
